+++ changelog.markdown
@@ -2,8 +2,7 @@
 
 * Bundled with docu.exe compiled against .Net 4.0.
 * Fixed docu calls to run with full filenames.
-* Added targetplatform switch for ILMerge task.
-* Added log switch for ILMerge task.
+* Added targetplatform, target and log switches for ILMerge task.
 * Added Git.Information.getLastTag() which gets the last git tag by calling git describe.
 * Added Git.Information.getCurrentHash() which gets the last current sha1.